PB-368 : create header module for everything header
searchbar, Swiss logo, etc...
