+++ src/modules/header/components/HeaderLink.vue
@@ -38,7 +38,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import 'src/scss/webmapviewer-bootstrap-theme';
+@import '@/scss/webmapviewer-bootstrap-theme';
 
 .custom-text-decoration {
     text-decoration: none;
